chore: order history e2e viewport support
