+++ projects/storefrontapp-e2e-cypress/cypress/helpers/language.ts
@@ -1,11 +1,6 @@
-import { clickHamburger } from './homepage';
-
 // Use language switcher to change language
 
-export function switchLanguage(lang: string, isMobile?: boolean) {
-  if (isMobile) {
-    clickHamburger();
-  }
+export function switchLanguage(lang: string) {
 
   cy.get('.SiteContext label')
     .contains('Language')
