+++ projects/storefrontapp-e2e-cypress/cypress/helpers/order-history.ts
@@ -5,7 +5,7 @@ import {
   replenishmentOrderHistoryUrl,
 } from './b2b/b2b-replenishment-order-history';
 import { waitForPage } from './checkout-flow';
-import { checkBanner } from './homepage';
+import { checkBanner, clickHamburger } from './homepage';
 import { switchLanguage } from './language';
 
 const orderHistoryLink = '/my-account/orders';
@@ -96,11 +96,10 @@ export const orderHistoryTest = {
   },
   checkSortingByCode() {
     it('should sort the orders table by given code', () => {
-      cy.server();
-      cy.route('GET', /sort=byOrderNumber/).as('query_order_asc');
+      cy.intercept('GET', /sort=byOrderNumber/).as('query_order_asc');
       cy.visit('/my-account/orders');
       cy.get('.top cx-sorting .ng-select').ngSelect('Order Number');
-      cy.wait('@query_order_asc').its('status').should('eq', 200);
+      cy.wait('@query_order_asc').its('response.statusCode').should('eq', 200);
       cy.get('.cx-order-history-code > .cx-order-history-value').then(
         ($orders) => {
           expect(parseInt($orders[0].textContent, 10)).to.be.lessThan(
@@ -110,15 +109,11 @@ export const orderHistoryTest = {
       );
     });
   },
-  checkCorrectDateFormat(isMobile?: boolean) {
+  checkCorrectDateFormat() {
     it('should show correct date format', () => {
-      cy.server();
-      cy.route(
-        'GET',
-        `${Cypress.env('OCC_PREFIX')}/${Cypress.env(
-          'BASE_SITE'
-        )}/cms/pages?*/my-account/orders*`
-      ).as('getOrderHistoryPage');
+      cy.intercept('GET', '/users/current/orders').as('getOrderHistoryPage');
+
+      cy.visit('/my-account/orders');
 
       // to compare two dates (EN and DE) we have to compare day numbers
       // EN: "June 15, 2019"
@@ -127,9 +122,13 @@ export const orderHistoryTest = {
       const getDayNumber = (element: any) =>
         element.text().replace(',', '').replace('.', '').split(' ');
       let dayNumberEN: string;
-      cy.visit('/my-account/orders');
-      cy.wait('@getOrderHistoryPage').its('status').should('eq', 200);
-      switchLanguage('en', isMobile);
+
+      cy.wait('@getOrderHistoryPage').its('response.statusCode').should('eq', 200);
+
+      cy.onMobile(() => {
+          clickHamburger();
+        });
+      switchLanguage('en');
 
       cy.get('.cx-order-history-placed > .cx-order-history-value')
         .first()
@@ -137,7 +136,10 @@ export const orderHistoryTest = {
           dayNumberEN = getDayNumber(element)[1];
         });
 
-      switchLanguage('de', isMobile);
+      cy.onMobile(() => {
+          clickHamburger();
+        });
+      switchLanguage('de');
 
       cy.get('.cx-order-history-placed > .cx-order-history-value')
         .first()
@@ -145,7 +147,10 @@ export const orderHistoryTest = {
           expect(getDayNumber(element)[0]).to.eq(dayNumberEN);
         });
 
-      switchLanguage('en', isMobile); // switch language back
+      cy.onMobile(() => {
+          clickHamburger();
+        });
+      switchLanguage('en'); // switch language back
     });
   },
 };
